ファイルをコンバートする処理を別クラスへ移動
